binance-trading-agent/requirements.txt
python-dotenv==1.0.1
requests==2.31.0

# Google AI
google-generativeai==0.8.5

# LangChain - Let pip resolve compatible versions
langchain==0.3.14
langchain-google-genai==2.0.10

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

